Rename robot controllers and add one for the Right Shadow hand.
In order to work with this controller (Shadow), we need a roscore running with the Aurobot and another roscore running with the Shadow driver. Using the multimaster_fkie package, the second roscore must synchronise the topic /rh_trajectory_controller/command, so the driver can send the positions generated by MoveIt (roscore with the Aurobot) to the actual robot.